requirements.txt
paho-mqtt==1.5.0
psycopg==3.2.1
psycopg-binary==3.2.1
python-dotenv~=1.0.0
schedule==1.2.2
numpy~=1.26.1
jsonpickle==3.2.2